Move `checkSession` to `common/external`
